+++ src/routes/submission/[id]/+page.svelte
@@ -9,9 +9,10 @@
 	import EditableText from '$lib/components/EditableText.svelte';
 	import { getDB } from '$lib/data/CRUD';
 	import { getAuth } from '../../Auth.svelte';
+	import Link from '$lib/components/Link.svelte';
 
 	let { data }: { data: PageData } = $props();
-	const { submission, venue, authors } = $derived(data);
+	const { submission, venue, authors, previous } = $derived(data);
 
 	const db = getDB();
 	const auth = getAuth();
@@ -46,8 +47,9 @@
 	>
 		{#snippet subtitle()}Submission{/snippet}
 		{#snippet details()}
-			{#if submission.previousid}{submission.previousid} →
-			{/if}{submission.externalid}
+			{#if previous}<Link to="/submission/{previous.id}">{previous.externalid}</Link>
+				→{/if}
+			{submission.externalid}
 		{/snippet}
 
 		<h2>Authors</h2>